purchase-backend/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Pragyo Tandukar <84461389+tandukar@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
psycopg2 = "^2.9.9"
celery = "^5.3.6"
django-split-settings = "^1.3.1"
pywhatkit = "^5.4"
vonage = "^3.13.0"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
mongoengine = "^0.28.2"
pymongo = {extras = ["gssapi", "snappy", "srv", "tls"], version = "^4.7.1"}
dnspython = "^2.6.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
pre-commit = "^3.7.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
